+++ Runtime/RealtimeDatabase/RealtimeDatabase.cs
@@ -20,57 +20,92 @@ namespace FirebaseRestClient
         private string endAtValue;
         private string equalToValue;
 
+        /// <summary>
+        /// Create Firebase Realtime Database Instance
+        /// </summary>
         public RealtimeDatabase() 
         {
             FirebaseSettings.LoadSettings();
         }
 
+        /// <summary>
+        /// Internal Constructor
+        /// </summary>
+        /// <param name="path">Previous Path</param>
         internal RealtimeDatabase(string path)
         {
             this.path = path;
         }
 
+        /// <summary>
+        /// Internal Constructor
+        /// </summary>
+        /// <param name="path">Previous Path</param>
+        /// <param name="filters">Filters</param>
         internal RealtimeDatabase(string path, FirebaseFilters filters)
         {
             this.path = path;
             FromFirebaseFilters(filters);
         }
 
+        /// <summary>
+        /// Point to Child Node
+        /// </summary>
+        /// <param name="value">Path</param>
+        /// <returns></returns>
         public RealtimeDatabase Child(string value)
         {
             string newPath = String.IsNullOrEmpty(path) ? value : $"{path}/{value}";
             return new RealtimeDatabase(newPath, ToFirebaseFilters());
         }
 
-        public GeneralCallback Write<T>(T body)
+        /// <summary>
+        /// Write Key-Value Pair to specified node
+        /// </summary>
+        /// <param name="key">Key Name</param>
+        /// <param name="value">Value, any datatypes</param>
+        /// <param name="append">Should append to existing list or just overwrite?</param>
+        /// <returns></returns>
+        public GeneralCallback WriteKeyValuePair(string key, object value, bool append = false)
         {
             GeneralCallback callbackHandler = new GeneralCallback();
+
+            string json = GenerateJson(value, key);
 
             RequestHelper req = new RequestHelper
             {
                 Uri = FirebaseConfig.endpoint + path + ".json" + GetAuthParam(),
-                Body = body
+                BodyString = json
             };
 
-            RESTHelper.Put<T>(req, res =>
+            if (append) //patch
             {
-                callbackHandler.successCallback?.Invoke();
-            },
-            err =>
-            {
-                callbackHandler.exceptionCallback?.Invoke(err);
-            });
+                RESTHelper.Patch(req, res => callbackHandler.successCallback?.Invoke(), 
+                    err => callbackHandler.exceptionCallback?.Invoke(err));
+                return callbackHandler;
+            }
+
+            RESTHelper.PutJson(req, res => callbackHandler.successCallback?.Invoke(), 
+                err => callbackHandler.exceptionCallback?.Invoke(err));
 
             return callbackHandler;
         }
 
-        public GeneralCallback Write(string json)
+        /// <summary>
+        /// Write to specified node
+        /// </summary>
+        /// <param name="value">Value</param>
+        /// <returns></returns>
+        public GeneralCallback WriteValue(object value)
         {
             GeneralCallback callbackHandler = new GeneralCallback();
 
+            string json = GenerateJson(value);
+            string upperNodePath = GetImmediateUpperNode();
+            
             RequestHelper req = new RequestHelper
             {
-                Uri = FirebaseConfig.endpoint + path + ".json" + GetAuthParam(),
+                Uri = FirebaseConfig.endpoint + upperNodePath + ".json" + GetAuthParam(),
                 BodyString = json
             };
 
@@ -86,68 +121,30 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
-        public GeneralCallback Write(string key, string value)
-        {
-            GeneralCallback callbackHandler = new GeneralCallback();
-
-            string json = "{" +
-                $"\"{key}\":\"{value}\"" +
-                "}";
-
-            RequestHelper req = new RequestHelper
-            {
-                Uri = FirebaseConfig.endpoint + path + ".json" + GetAuthParam(),
-                BodyString = json
-            };
-
-            RESTHelper.PutJson(req, res =>
-            {
-                callbackHandler.successCallback?.Invoke();
-            },
-            err =>
-            {
-                callbackHandler.exceptionCallback?.Invoke(err);
-            });
-
-            return callbackHandler;
-        }
-
-        public GeneralCallback WriteAppend(string key, string value)
-        {
-            GeneralCallback callbackHandler = new GeneralCallback();
-
-            string json = "{" +
-                $"\"{key}\":\"{value}\"" +
-                "}";
-
-            RequestHelper req = new RequestHelper
-            {
-                Uri = FirebaseConfig.endpoint + path + ".json" + GetAuthParam(),
-                BodyString = json
-            };
-
-            RESTHelper.Patch(req, res =>
-            {
-                callbackHandler.successCallback?.Invoke();
-            },
-            err =>
-            {
-                callbackHandler.exceptionCallback?.Invoke(err);
-            });
-
-            return callbackHandler;
-        }
-
+        /// <summary>
+        /// Generate Firebase Push ID
+        /// </summary>
+        /// <returns>Unique Push ID</returns>
         public string GeneratePushID()
         {
             return PushIDGenerator.GeneratePushID();
         }
 
+        /// <summary>
+        /// Convert Push ID
+        /// </summary>
+        /// <param name="pushId">Push ID</param>
+        /// <returns></returns>
         public long ConvertPushID(string pushId)
         {
             return PushIDGenerator.ConvertPushID(pushId);   
         }
 
+        /// <summary>
+        /// Push Raw Json to specified node
+        /// </summary>
+        /// <param name="json">Valid Raw Json</param>
+        /// <returns>Generated Push ID</returns>
         public StringCallback Push(string json)
         {
             StringCallback callbackHandler = new StringCallback();
@@ -172,6 +169,12 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
+        /// <summary>
+        /// Push object as Body
+        /// </summary>
+        /// <typeparam name="T">Type of Body</typeparam>
+        /// <param name="body">Body</param>
+        /// <returns>Generated Push ID</returns>
         public StringCallback Push<T>(T body)
         {
             StringCallback callbackHandler = new StringCallback();
@@ -196,6 +199,10 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
+        /// <summary>
+        /// Remove value of specified Node
+        /// </summary>
+        /// <returns></returns>
         public GeneralCallback Remove()
         {
             GeneralCallback callbackHandler = new GeneralCallback();
@@ -214,31 +221,20 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
-
-        public DictionaryCallback<string, string> Read()
+        /// <summary>
+        /// Read Json Object
+        /// </summary>
+        /// <typeparam name="T">Desired type to be converted</typeparam>
+        /// <returns>Result in object format</returns>
+        public ObjectCallback<T> Read<T>()
         {
-            DictionaryCallback<string, string> callbackHandler = new DictionaryCallback<string, string>();
+            ObjectCallback<T> callbackHandler = new ObjectCallback<T>();
 
             string route = FirebaseConfig.endpoint + path + ".json" + GetAuthParam();
 
             RESTHelper.Get(route, res =>
             {
-                var resData = fsJsonParser.Parse(res.Text); //in JSON 
-                Dictionary<string, string> destructuredRes = new Dictionary<string, string>();
-
-                if (resData.IsDictionary)
-                {
-                    object deserializedRes = null;
-                    fsSerializer serializer = new fsSerializer();
-                    serializer.TryDeserialize(resData, typeof(Dictionary<string, string>), ref deserializedRes);
-                    destructuredRes = (Dictionary<string, string>)deserializedRes;
-                    callbackHandler.successCallback?.Invoke(destructuredRes);
-                    return;
-                }
-                //No collection, single result (key-value pair)
-                string[] splittedPaths = path.Split('/');
-                destructuredRes.Add(splittedPaths[splittedPaths.Length - 1], resData._value.ToString()); //last path as key, resData._value as value
-                callbackHandler.successCallback?.Invoke(destructuredRes);
+                callbackHandler.successCallback?.Invoke(JsonUtility.FromJson<T>(res.Text));
             },
             err =>
             {
@@ -248,33 +244,11 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
-        public DictionaryCallback<string, T> Read<T>()
-        {
-            DictionaryCallback<string, T> callbackHandler = new DictionaryCallback<string, T>();
-
-            string route = FirebaseConfig.endpoint + path + ".json" + GetAuthParam();
-
-            RESTHelper.Get(route, res =>
-            {
-                var resData = fsJsonParser.Parse(res.Text); 
-
-                object deserializedRes = null;
-
-                fsSerializer serializer = new fsSerializer();
-                serializer.TryDeserialize(resData, typeof(Dictionary<string, T>), ref deserializedRes);
-
-                Dictionary<string, T> destructuredRes = (Dictionary<string, T>)deserializedRes;
-
-                callbackHandler.successCallback?.Invoke(destructuredRes);
-            },
-            err =>
-            {
-                callbackHandler.exceptionCallback?.Invoke(err);
-            });
-
-            return callbackHandler;
-        }
-
+        /// <summary>
+        /// Read raw Json
+        /// </summary>
+        /// <param name="shallow">If shallow, all json object will return as true. Better for surface reading.</param>
+        /// <returns>Raw Json</returns>
         public StringCallback RawRead(bool shallow=false)
         {
             StringCallback callbackHandler = new StringCallback();
@@ -295,9 +269,13 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
-        public StringCallback ReadValue()
+        /// <summary>
+        /// Read value of key
+        /// </summary>
+        /// <returns>Json Object as Json String, else object</returns>
+        public ObjectCallback ReadValue()
         {
-            StringCallback callbackHandler = new StringCallback();
+            ObjectCallback callbackHandler = new ObjectCallback();
 
             string route = FirebaseConfig.endpoint + path + ".json" + GetAuthParam();
 
@@ -311,7 +289,7 @@ namespace FirebaseRestClient
                     return;
                 }
                 //No collection
-                string value = resData._value.ToString();
+                object value = resData._value;
                 callbackHandler.successCallback?.Invoke(value);
             },
             err =>
@@ -322,6 +300,77 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
+        /// <summary>
+        /// Read key-value pairs
+        /// </summary>
+        /// <returns>Dictionary, both key and value is object, it can be any primitative types or fsData</returns>
+        public DictionaryCallback<string, object> ReadKeyValuePairs()
+        {
+            DictionaryCallback<string, object> callbackHandler = new DictionaryCallback<string, object>();
+
+            string route = FirebaseConfig.endpoint + path + ".json" + GetAuthParam();
+
+            RESTHelper.Get(route, res =>
+            {
+                var resData = fsJsonParser.Parse(res.Text); //in JSON 
+                Dictionary<string, object> destructuredRes = new Dictionary<string, object>();
+
+                if (resData.IsDictionary)
+                {
+                    resData.AsDictionary.ToList().ForEach(x => destructuredRes.Add(x.Key, x.Value._value));
+                    callbackHandler.successCallback?.Invoke(destructuredRes);
+                    return;
+                }
+                //No collection, single result (key-value pair)
+                string[] splittedPaths = path.Split('/');
+                destructuredRes.Add(splittedPaths[splittedPaths.Length - 1], resData._value.ToString()); //last path as key, resData._value as value
+                callbackHandler.successCallback?.Invoke(destructuredRes);
+            },
+            err =>
+            {
+                callbackHandler.exceptionCallback?.Invoke(err);
+            });
+
+            return callbackHandler;
+        }
+
+        /// <summary>
+        /// Read key-value pairs
+        /// </summary>
+        /// <typeparam name="T">Value type</typeparam>
+        /// <returns>Dictionary, key as string, value as specified type. Returns nothing if failed to deserialize.</returns>
+        public DictionaryCallback<string, T> ReadKeyValuePairs<T>()
+        {
+            DictionaryCallback<string, T> callbackHandler = new DictionaryCallback<string, T>();
+
+            string route = FirebaseConfig.endpoint + path + ".json" + GetAuthParam();
+
+            RESTHelper.Get(route, res =>
+            {
+                var resData = fsJsonParser.Parse(res.Text);
+
+                object deserializedRes = null;
+
+                fsSerializer serializer = new fsSerializer();
+                serializer.TryDeserialize(resData, typeof(Dictionary<string, T>), ref deserializedRes);
+
+                Dictionary<string, T> destructuredRes = (Dictionary<string, T>)deserializedRes;
+
+                callbackHandler.successCallback?.Invoke(destructuredRes);
+            },
+            err =>
+            {
+                callbackHandler.exceptionCallback?.Invoke(err);
+            });
+
+            return callbackHandler;
+        }
+
+        /// <summary>
+        /// Check for child in selected path
+        /// </summary>
+        /// <param name="child">child name</param>
+        /// <returns>Returns true if child exists, else false</returns>
         public BooleanCallback HasChild(string child)
         {
             BooleanCallback callbackHandler = new BooleanCallback();
@@ -348,14 +397,23 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
-        public GeneralCallback Update<T>(T body)
+
+        /// <summary>
+        /// Update value of specified path
+        /// </summary>
+        /// <param name="value">Updated value</param>
+        /// <returns></returns>
+        public GeneralCallback Update(object value)
         {
             GeneralCallback callbackHandler = new GeneralCallback();
 
+            string json = GenerateJson(value);
+            string upperNodePath = GetImmediateUpperNode();
+
             RequestHelper req = new RequestHelper
             {
-                Uri = FirebaseConfig.endpoint + path + ".json" + GetAuthParam(),
-                Body = body
+                Uri = FirebaseConfig.endpoint + upperNodePath + ".json" + GetAuthParam(),
+                BodyString = json
             };
             RESTHelper.Patch(req, res =>
             {
@@ -369,36 +427,24 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
-        public GeneralCallback Update(string json)
-        {
-            GeneralCallback callbackHandler = new GeneralCallback();
-
-            RequestHelper req = new RequestHelper
-            {
-                Uri = FirebaseConfig.endpoint + path + ".json" + GetAuthParam(),
-                BodyString = json,
-            };
-            RESTHelper.Patch(req, res =>
-            {
-                callbackHandler.successCallback?.Invoke();
-            },
-            err =>
-            {
-                callbackHandler.exceptionCallback?.Invoke(err);
-            });
-
-            return callbackHandler;
-        }
-
-        public void ValueChanged(Action<ChildEventArgs> callback)
+        /// <summary>
+        /// Union of all server events - ChildAdded, ChildRemoved, ChildChanged
+        /// </summary>
+        /// <param name="callback"></param>
+        public void ValueChanged(Action<ChildEventArgs> callback, bool shallow = false)
         {
             ChildEventHandler childEventHandler = new ChildEventHandler();
             childEventHandler.OnChildEventReceived += callback;
             childEventHandler.childEventType = ChildEventHandler.ChildEventType.ValueChanged;
-
+            childEventHandler.shallow = shallow;
             ChildEventListen(childEventHandler);
         }
 
+        /// <summary>
+        /// Listen for Child addition to targeted path and in-depth.
+        /// </summary>
+        /// <param name="callback"></param>
+        /// <param name="shallow">Listen to surface level child, not in-depth child</param>
         public void ChildAdded(Action<ChildEventArgs> callback, bool shallow=false)
         {
             ChildEventHandler childEventHandler = new ChildEventHandler();
@@ -408,21 +454,29 @@ namespace FirebaseRestClient
             ChildEventListen(childEventHandler);
         }
 
-        public void ChildRemoved(Action<ChildEventArgs> callback)
+        /// <summary>
+        /// Listen for Child Remove at any level from targeted path.
+        /// </summary>
+        /// <param name="callback"></param>
+        public void ChildRemoved(Action<ChildEventArgs> callback, bool shallow = false)
         {
             ChildEventHandler childEventHandler = new ChildEventHandler();
             childEventHandler.OnChildEventReceived += callback;
             childEventHandler.childEventType = ChildEventHandler.ChildEventType.ChildRemoved;
-
+            childEventHandler.shallow = shallow;
             ChildEventListen(childEventHandler);
         }
 
-        public void ChildChanged(Action<ChildEventArgs> callback)
+        /// <summary>
+        /// Combination of both ChildAdded and ChildRemoved.
+        /// </summary>
+        /// <param name="callback"></param>
+        public void ChildChanged(Action<ChildEventArgs> callback, bool shallow = false)
         {
             ChildEventHandler childEventHandler = new ChildEventHandler();
             childEventHandler.OnChildEventReceived += callback;
             childEventHandler.childEventType = ChildEventHandler.ChildEventType.ChildChanged;
-
+            childEventHandler.shallow = shallow;
             ChildEventListen(childEventHandler);
         }
 
@@ -441,6 +495,7 @@ namespace FirebaseRestClient
                 RetrySecondsDelay = 1
             };
 
+            //create an unsubscriber for events
             var unsubscriberGO = GameObject.Find("FirebaseRestUnsubscriber");
             if (unsubscriberGO == null)
             {
@@ -452,6 +507,11 @@ namespace FirebaseRestClient
             RESTHelper.Get(req, err => RequestErrorHelper.ToDictionary(err).ToList().ForEach(x => Debug.Log(x.Key + " - " + x.Value)));
         }
 
+        /// <summary>
+        /// Order results by the value of a specified child key
+        /// </summary>
+        /// <param name="key">Child Key</param>
+        /// <returns>Raw Json</returns>
         public StringCallback OrderByChild(string key)
         {
             StringCallback callbackHandler = new StringCallback();
@@ -480,6 +540,10 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
+        /// <summary>
+        /// Order results by child keys. [Firebase Rules must be set]
+        /// </summary>
+        /// <returns>Raw Json</returns>
         public StringCallback OrderByKey()
         {
             StringCallback callbackHandler = new StringCallback();
@@ -508,6 +572,10 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
+        /// <summary>
+        /// Order results by child value. [Firebase rules must be set]
+        /// </summary>
+        /// <returns>Raw Json</returns>
         public StringCallback OrderByValue()
         {
             StringCallback callbackHandler = new StringCallback();
@@ -536,37 +604,69 @@ namespace FirebaseRestClient
             return callbackHandler;
         }
 
-
-        public RealtimeDatabase LimitToFirst(string value)
+        /// <summary>
+        /// Maximum numbers to return from the beginning
+        /// </summary>
+        /// <param name="value">Maximum return</param>
+        /// <returns></returns>
+        public RealtimeDatabase LimitToFirst(int? value)
         {
-            limitToFirstValue = value;
+            limitToFirstValue = value.ToString();
             return new RealtimeDatabase(path, ToFirebaseFilters());
         }
 
-        public RealtimeDatabase LimitToLast(string value)
+        /// <summary>
+        /// Maximum numbers to return from the end
+        /// </summary>
+        /// <param name="value"></param>
+        /// <returns></returns>
+        public RealtimeDatabase LimitToLast(int? value)
         {
-            limitToLastValue = value;
+            limitToLastValue = value.ToString();
             return new RealtimeDatabase(path, ToFirebaseFilters());
         }
 
-        public RealtimeDatabase StartAt(string value)
+        /// <summary>
+        /// Returns items greater than or equal to specified value
+        /// </summary>
+        /// <param name="value">Filter Value</param>
+        /// <returns></returns>
+        public RealtimeDatabase StartAt(object value)
         {
-            startAtValue = value;
+            if (value != null) startAtValue = value.GetType() == typeof(string) && !String.IsNullOrEmpty(value.ToString()) ? $"\"{value}\"" : value.ToString(); //append quote if it's string
+            else startAtValue = "";
             return new RealtimeDatabase(path, ToFirebaseFilters());
         }
 
-        public RealtimeDatabase EndAt(string value)
+        /// <summary>
+        /// Returns items less than or equal specified value
+        /// </summary>
+        /// <param name="value">Filter Value</param>
+        /// <returns></returns>
+        public RealtimeDatabase EndAt(object value)
         {
-            endAtValue = value;
+            if (value != null) endAtValue = value.GetType() == typeof(string) && !String.IsNullOrEmpty(value.ToString()) ? $"\"{value}\"" : value.ToString(); //append quote if it's string
+            else endAtValue = "";
             return new RealtimeDatabase(path, ToFirebaseFilters());
         }
 
-        public RealtimeDatabase EqualTo(string value)
+        /// <summary>
+        /// Returns items equal to specified value
+        /// </summary>
+        /// <param name="value">Filter Value</param>
+        /// <returns></returns>
+        public RealtimeDatabase EqualTo(object value)
         {
-            equalToValue = value;
+            if (value != null) equalToValue = value.GetType() == typeof(string) && !String.IsNullOrEmpty(value.ToString()) ? $"\"{value}\"" : value.ToString(); //append quote if it's string
+            else equalToValue = "";
             return new RealtimeDatabase(path, ToFirebaseFilters());
         }
 
+        /// <summary>
+        /// Generate filters for query
+        /// </summary>
+        /// <param name="appendQuote">Surround string with quote</param>
+        /// <returns></returns>
         Dictionary<string, string> GetFilterCollection(bool appendQuote)
         {
             Dictionary<string, string> filters = new Dictionary<string, string>();
@@ -589,6 +689,10 @@ namespace FirebaseRestClient
             return filters;
         }
 
+        /// <summary>
+        /// Convert props to FirebaseFilters
+        /// </summary>
+        /// <returns></returns>
         FirebaseFilters ToFirebaseFilters()
         {
             return new FirebaseFilters
@@ -601,6 +705,10 @@ namespace FirebaseRestClient
             };
         }
 
+        /// <summary>
+        /// Assign to props from FirebaseFilters
+        /// </summary>
+        /// <param name="filters"></param>
         void FromFirebaseFilters(FirebaseFilters filters)
         {
             limitToFirstValue = filters.limitToFirstValue;
@@ -626,6 +734,76 @@ namespace FirebaseRestClient
         string GetShallowParam()
         {
             return FirebaseAuthentication.currentUser != null ? "" : $"?shallow=true";
+        }
+
+        /// <summary>
+        /// Generate server friendly Json
+        /// </summary>
+        /// <param name="value">Any datatypes</param>
+        /// <param name="key">Key Value</param>
+        /// <returns></returns>
+        string GenerateJson(object value, string key = "")
+        {
+            //Get key
+            if (String.IsNullOrEmpty(key))
+            {
+                string[] splittedPaths = path.Split('/');
+                key = splittedPaths[splittedPaths.Length - 1];
+            }
+
+            string json = "{" +
+            $"\"{key}\":{value}" +
+            "}";
+
+            var paramType = value.GetType();
+            //If param is not primitive, convert to json
+            if (!paramType.IsPrimitive && paramType != typeof(String))
+            {
+                string paramJson = JsonUtility.ToJson(value);
+                json = "{" +
+                $"\"{key}\":{paramJson}" +
+                "}";
+                return json;
+            }
+            //if normal string, wrap with quote
+            if (paramType == typeof(String))
+            {
+                string paramString = value.ToString();
+                int length = paramString.Length;
+                //make sure it's not json, if json, leave as it is.
+                if (paramString[0] != '{' && paramString[length - 1] != '}')
+                {
+                    json = "{" +
+                    $"\"{key}\":\"{value}\"" +
+                    "}";
+                    return json;
+                }
+            }
+            //bool type should to lowercase to work in json
+            if (paramType == typeof(bool))
+            {
+                json = "{" +
+                $"\"{key}\":{value.ToString().ToLower()}" +
+                "}";
+            }
+            return json;
+        }
+
+        /// <summary>
+        /// Returns upper node from passed node arg
+        /// </summary>
+        /// <returns></returns>
+        string GetImmediateUpperNode()
+        {
+            //Create immediate upper node path
+            string[] splittedPaths = path.Split('/');
+            Array.Resize(ref splittedPaths, splittedPaths.Length - 1);
+            string upperNodePath = ""; //one level up path
+            splittedPaths.ToList().
+                ForEach(s => upperNodePath = String.IsNullOrEmpty(upperNodePath) ?
+                upperNodePath += s : upperNodePath += "/" + s);
+
+            return upperNodePath;
         }
     }